tests/wpt/web-platform-tests/html/semantics/scripting-1/the-script-element/module/dynamic-import/dynamic-imports-error.html

<!DOCTYPE html>
<title>Basic dynamic imports</title>
<link rel="author" title="Kouhei Ueno" href="mailto:kouhei@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script type="module">
promise_test(t => {
  return promise_rejects(t, new SyntaxError, import("../syntaxerror.js"));
}, "Dynamic imports should reject if the imported tree has a syntax error.");
promise_test(t => {
  return promise_rejects(t, new SyntaxError, import("../instantiation-error-1.js"));
}, "Dynamic imports should reject if the imported tree has an instantiation error.");
promise_test(t => {
  return promise_rejects(t, new Error, import("../throw-error.js")).then(() => {
    assert_true(window.before_throwing_error,
        "the module script should run to a point where it throws exception");
    assert_equals(undefined, window.after_throwing_error,
        "the module script should not run after it throws exception");
  });
}, "Dynamic imports should reject if the imported tree has an evaluation error.");
</script>
